+++ src/main/java/roomescape/domain/ReservationDto.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.domain;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
